sensor: unify struct regval_list placement and layout
Every sensor driver defined 'struct regval_list' with one of eight
different bodies and in one of several locations (under SPECIAL
FEATURES, after module_param/sensor_info, or before the register
tables). Normalise to a single form:

    struct regval_list {
            uint16_t reg_num;
            uint16_t value;
    };

uint16_t for both fields is a safe superset: no register value in any
driver exceeds 0xffff, and every sensor_read/sensor_write accepts a
16-bit register argument (8-bit drivers truncate at the call, as they
already did). This also flattens the '#ifdef SENSOR_I2C_REG_8BIT/16BIT'
bodies to one definition.

Placement: immediately after the sensor_info declaration where present,
otherwise immediately before struct again_lut (i.e. the top of the
declarations), so the type is always defined before its first use.

Two pre-existing bugs surfaced and are fixed here:
  - 3.10.14/t41/imx415.c used struct regval_list but never defined it
  - 3.10.14/t41/rn6752.c defined it after the tables that use it

Also folds in a user fix: 4.4.94/t31/jxq03p.c used sizeof("jxq03")
instead of sizeof(SENSOR_NAME) in memcpy(chip->name, ...).

923 files. Verified: T31/3.10.14 (22 sensors), T31/4.4.94 (16),
t23/3.10.14 (8, incl. os02n10 with the flattened #ifdef body).

+++ 4.4.94/sensor-src/t31/mis2006.c
@@ -83,8 +83,8 @@ static struct sensor_info sensor_info = {
 };
 
 struct regval_list {
-	unsigned int reg_num;
-	unsigned int value;
+	uint16_t reg_num;
+	uint16_t value;
 };
 
 struct again_lut {
