+++ 4.4.94/sensor-src/t31/mis2008.c
@@ -87,8 +87,8 @@ static struct sensor_info sensor_info = {
 };
 
 struct regval_list {
-	unsigned int reg_num;
-	unsigned int value;
+	uint16_t reg_num;
+	uint16_t value;
 };
 
 struct again_lut {
