+++ 4.4.94/sensor-src/t40/mis2006.c
@@ -64,8 +64,8 @@ module_param(sensor_max_fps, int, S_IRUGO);
 MODULE_PARM_DESC(sensor_max_fps, "Sensor Max Fps set interface");
 
 struct regval_list {
-	unsigned int reg_num;
-	unsigned int value;
+	uint16_t reg_num;
+	uint16_t value;
 };
 
 struct again_lut {
