+++ common/sensor-src/t23/mis2009.c
@@ -70,9 +70,10 @@ MODULE_PARM_DESC(sensor_max_fps, "Sensor Max Fps set interface");
 static int shvflip = 1;
 module_param(shvflip, int, S_IRUGO);
 MODULE_PARM_DESC(shvflip, "Sensor HV Flip Enable interface");
+
 struct regval_list {
-	unsigned int reg_num;
-	unsigned int value;
+	uint16_t reg_num;
+	uint16_t value;
 };
 
 struct again_lut {
