+++ common/sensor-src/t23/mis20c1.c
@@ -81,8 +81,8 @@ static struct sensor_info sensor_info = {
 };
 
 struct regval_list {
-	u16 reg_num;
-	u8 value;
+	uint16_t reg_num;
+	uint16_t value;
 };
 
 struct again_lut {
